+++ src/main/java/com/lowdragmc/lowdraglib2/nodegraphtookit/gui/command/NodeCommands.java
@@ -303,9 +303,5 @@ public final class NodeCommands {
             creationData.clear();
         }
 
-        @Override
-        public void undo() {
-
-        }
     }
 }