+++ src/main/java/com/lowdragmc/lowdraglib2/nodegraphtookit/model/DeclarationModel.java
@@ -1,5 +1,6 @@
 package com.lowdragmc.lowdraglib2.nodegraphtookit.model;
 
+import com.lowdragmc.lowdraglib2.syncdata.annotation.Persisted;
 import lombok.Getter;
 import lombok.Setter;
 import net.minecraft.network.chat.Component;
@@ -10,7 +11,7 @@ import java.util.List;
  * A model that represents a declaration (e.g. a variable) in a graph.
  */
 public class DeclarationModel extends GraphElementModel implements IHasName {
-    @Getter
+    @Persisted @Getter
     public String name = "";
 
     public DeclarationModel() {